rename utilities and update docs
